+++ lib/Model/MerchantReturnRequest.php
@@ -64,8 +64,8 @@ class MerchantReturnRequest implements ModelInterface, ArrayAccess
         'reason' => 'string',
         'customerComment' => 'string',
         'merchantComment' => 'string',
-        'refundInclVat' => 'double',
-        'refundExclVat' => 'double'
+        'refundInclVat' => 'float',
+        'refundExclVat' => 'float'
     ];
 
     /**
@@ -81,8 +81,8 @@ class MerchantReturnRequest implements ModelInterface, ArrayAccess
         'reason' => null,
         'customerComment' => null,
         'merchantComment' => null,
-        'refundInclVat' => 'double',
-        'refundExclVat' => 'double'
+        'refundInclVat' => 'decimal',
+        'refundExclVat' => 'decimal'
     ];
 
     /**
@@ -534,7 +534,7 @@ class MerchantReturnRequest implements ModelInterface, ArrayAccess
     /**
      * Gets refundInclVat
      *
-     * @return double
+     * @return float
      */
     public function getRefundInclVat()
     {
@@ -544,7 +544,7 @@ class MerchantReturnRequest implements ModelInterface, ArrayAccess
     /**
      * Sets refundInclVat
      *
-     * @param double $refundInclVat Refund amount incl. VAT
+     * @param float $refundInclVat Refund amount incl. VAT
      *
      * @return $this
      */
@@ -558,7 +558,7 @@ class MerchantReturnRequest implements ModelInterface, ArrayAccess
     /**
      * Gets refundExclVat
      *
-     * @return double
+     * @return float
      */
     public function getRefundExclVat()
     {
@@ -568,7 +568,7 @@ class MerchantReturnRequest implements ModelInterface, ArrayAccess
     /**
      * Sets refundExclVat
      *
-     * @param double $refundExclVat Refund amount excl. VAT
+     * @param float $refundExclVat Refund amount excl. VAT
      *
      * @return $this
      */
